create Libc and Libdl modules for the relevant functions
