+++ base/dates/conversions.jl
@@ -18,7 +18,7 @@ end
 # Returns unix seconds since 1970-01-01T00:00:00
 datetime2unix(dt::DateTime) = (value(dt) - UNIXEPOCH)/1000.0
 function now()
-    tm = TmStruct(time())
+    tm = Libc.TmStruct(time())
     return DateTime(tm.year+1900,tm.month+1,tm.mday,tm.hour,tm.min,tm.sec)
 end
 today() = Date(now())
