+++ base/exports.jl
@@ -9,6 +9,8 @@ export
     Dates,
     Sys,
     Test,
+    Libc,
+    Libdl,
     LinAlg,
     BLAS,
     LAPACK,
@@ -34,7 +36,6 @@ export
     BufferStream,
     CartesianIndex,
     CartesianRange,
-    CFILE,
     Cmd,
     Colon,
     Complex,
@@ -108,7 +109,6 @@ export
     Symmetric,
     SymTridiagonal,
     Timer,
-    TmStruct,
     Tridiagonal,
     UnitRange,
     UpperTriangular,
@@ -164,7 +164,6 @@ export
     ARGS,
     C_NULL,
     CPU_CORES,
-    DL_LOAD_PATH,
     ENDIAN_BOM,
     ENV,
     Inf,
@@ -172,21 +171,10 @@ export
     Inf32,
     JULIA_HOME,
     LOAD_PATH,
-    MS_ASYNC,
-    MS_INVALIDATE,
-    MS_SYNC,
     NaN,
     NaN16,
     NaN32,
     OS_NAME,
-    RTLD_DEEPBIND,
-    RTLD_FIRST,
-    RTLD_GLOBAL,
-    RTLD_LAZY,
-    RTLD_LOCAL,
-    RTLD_NODELETE,
-    RTLD_NOLOAD,
-    RTLD_NOW,
     STDERR,
     STDIN,
     STDOUT,
@@ -1022,8 +1010,6 @@ export
 
 # time
     sleep,
-    strftime,
-    strptime,
     tic,
     time,
     time_ns,
@@ -1132,7 +1118,6 @@ export
     fd,
     fdio,
     flush,
-    flush_cstdio,
     getaddrinfo,
     gethostname,
     getipaddr,
@@ -1145,10 +1130,8 @@ export
     listenany,
     ltoh,
     mark,
-    mmap,
     mmap_array,
     mmap_bitarray,
-    msync,
     nb_available,
     ntoh,
     open,
@@ -1317,25 +1300,13 @@ export
     success,
 
 # C interface
-    c_free,
-    c_malloc,
-    c_calloc,
-    c_realloc,
     cfunction,
     cglobal,
     disable_sigint,
-    dlclose,
-    dlopen,
-    dlopen_e,
-    dlsym,
-    dlsym_e,
-    errno,
-    find_library,
     pointer,
     pointer_from_objref,
     pointer_to_array,
     reenable_sigint,
-    strerror,
     unsafe_copy!,
     unsafe_load,
     unsafe_pointer_to_objref,
