+++ base/precompile.jl
@@ -151,7 +151,7 @@ precompile(Base.Terminals.TTYTerminal, (ASCIIString, Base.TTY, Base.TTY, Base.TT
 precompile(Base.Terminals.beep, (Base.Terminals.TTYTerminal,))
 precompile(Base.Terminals.raw!, (Base.Terminals.TTYTerminal, Bool))
 precompile(Base.Terminals.write, (Base.Terminals.TTYTerminal, Array{UInt8, 1}))
-precompile(Base.TmStruct, (Float64,))
+precompile(Libc.TmStruct, (Float64,))
 precompile(Base.VersionNumber, (Int, Int, Int, (), (ASCIIString,)))
 precompile(Base._atexit, ())
 precompile(Base._deleteat!, (Array{UInt8, 1}, Int, Int))
@@ -405,7 +405,7 @@ precompile(Base.start, (Dict{Symbol,Any},))
 precompile(Base.start, (UnitRange{Int},))
 precompile(Base.start_reading, (Base.TTY,))
 precompile(Base.stop_reading, (Base.TTY,))
-precompile(Base.strftime, (ASCIIString, Base.TmStruct))
+precompile(Libc.strftime, (ASCIIString, Libc.TmStruct))
 precompile(Base.string, (Int,))
 precompile(Base.strip, (ASCIIString,))
 precompile(Base.strwidth, (ASCIIString,))
