+++ test/file.jl
@@ -194,7 +194,7 @@ s = open(file, "r+")
 @test isreadonly(s) == false
 c = mmap_array(UInt8, (11,), s)
 c[5] = UInt8('x')
-msync(c)
+Libc.msync(c)
 close(s)
 s = open(file, "r")
 str = readline(s)
@@ -217,7 +217,7 @@ close(s)
 s = open(file, "r+")
 b = mmap_bitarray((17,19), s)
 rand!(b)
-msync(b)
+Libc.msync(b)
 b0 = copy(b)
 close(s)
 s = open(file, "r")
@@ -390,7 +390,7 @@ f = open(file, "w")
 write(f, "Hello, world!")
 close(f)
 f = open(file, "r")
-FILEp = convert(CFILE, f)
+FILEp = convert(Libc.FILE, f)
 buf = Array(UInt8, 8)
 str = ccall(:fread, Csize_t, (Ptr{Void}, Csize_t, Csize_t, Ptr{Void}), buf, 1, 8, FILEp.ptr)
 @test bytestring(buf) == "Hello, w"
